+++ unreal_plugin/Source/UnrealMCP/Public/UnrealMCPModule.h
@@ -27,4 +27,4 @@ public:
 	{
 		return FModuleManager::Get().IsModuleLoaded("UnrealMCP");
 	}
-}; 
+};